In progress debugging
At this point we get this error:

```
0=Unexpected value 'SkyAppRuntimeModule in
/Users/bradparker/code/blackbaud/skyux-builder/.e2e-tmp/node_modules/@blackbaud/skyux-builder/runtime/runtime.module.ts'
imported by the module 'SkyPagesModule in
/Users/bradparker/code/blackbaud/skyux-builder/.e2e-tmp/.skypagestmp/src/app/sky-pages.module.ts'.
Please add a @NgModule annotation.,
1=./.skypagestmp/src/main-internal.aot.ts
```

This isn't surprising as I've removed the NgModule annotation. I've done
this because it appears that any Angular require within the `./runtime`
dir triggers this error:

```
0=Error encountered resolving symbol values statically. Calling function
'ɵmakeDecorator', function calls are not supported. Consider replacing the
function or lambda with a reference to an exported function, resolving symbol
Injectable in
/Users/bradparker/code/blackbaud/skyux-builder/.e2e-tmp/node_modules/@blackbaud/skyux-builder/node_modules/@angular/core/core.d.ts,
resolving symbol ɵe in
/Users/bradparker/code/blackbaud/skyux-builder/.e2e-tmp/node_modules/@blackbaud/skyux-builder/node_modules/@angular/core/core.d.ts,
resolving symbol ɵe in
/Users/bradparker/code/blackbaud/skyux-builder/.e2e-tmp/node_modules/@blackbaud/skyux-builder/node_modules/@angular/core/core.d.ts,
1=./.skypagestmp/src/main-internal.aot.ts
```

For example if I make the following change:

```
--- a/runtime/auth-http.ts
+++ b/runtime/auth-http.ts
@@ -1,6 +1,6 @@
-// import {
-//   Injectable
-// } from '@angular/core';
+import {
+  Injectable
+} from '@angular/core';
```

This is enough to trigger the above error. This is true for _any_
angular import within the `./runtime` dir.